implement api call to next-webapp for video processing annd complete

+++ server.js
@@ -1,17 +1,17 @@
 const express = require("express")
 const cors = require("cors")
-const {Server} = require('socket.io')
+const { Server } = require('socket.io')
 const fs = require("fs")
 const http = require("http")
 const dotenv = require("dotenv")
-const {Readable} = require("stream")
+const { Readable } = require("stream")
 const axios = require("axios")
 // const {S3Client, PutObjectCommand} = require("@aws-sdk/client-s3")
 // const OpenAI = require("openai")
 const cloudinary = require("cloudinary").v2
 
 const app = express()
-const server =http.createServer(app)
+const server = http.createServer(app)
 app.use(cors())
 dotenv.config()
 
@@ -33,22 +33,22 @@ cloudinary.config({
 //     region: process.env.BUCKET_REGION
 // })
 
-const io = new Server(server,{
+const io = new Server(server, {
     cors: {
         origin: process.env.ELECTRON_HOST,
-        methods: ["GET","POST"],
+        methods: ["GET", "POST"],
     },
 })
 
 let recordedChunks = [];
 
-io.on("connection",(socket) => {
+io.on("connection", (socket) => {
     console.log("🟢 Socket is connected")
     socket.on("video-chunks", async (data) => {
-        console.log("🟢 Video chunk is sent",data)
-        const writeStream = fs.createWriteStream("temp_upload/"+data.filename)
+        console.log("🟢 Video chunk is sent", data)
+        const writeStream = fs.createWriteStream("temp_upload/" + data.filename)
         recordedChunks.push(data.chunks)
-        const videoBlob = new Blob(recordedChunks,{
+        const videoBlob = new Blob(recordedChunks, {
             type: "video/webm; codecs=vp9",
         })
         const buffer = Buffer.from(await videoBlob.arrayBuffer())
@@ -58,32 +58,44 @@ io.on("connection",(socket) => {
         })
 
     })
-    socket.on("process-video",async (data) => {
-        console.log("🟢 Video is processing",data)
+    socket.on("process-video", async (data) => {
+        console.log("🟢 Video is processing", data)
         recordedChunks = []
-        fs.readFile("temp_upload/"+data.filename, async (err,file) => {
+        fs.readFile("temp_upload/" + data.filename, async (err, file) => {
             if (err) {
                 console.log("🔴 Error reading file", err);
                 return;
             }
 
+            const processing = await axios.post(`${process.env.NEXT_API_HOST}recording/${data.userId}/processing`,{
+                filename: data.filename
+            })
+            if (processing.data.status !== 200) return console.log("🔴 Error: Something went wrong with creating the processing file")
+
             const cloudinaryUpload = await cloudinary.uploader.upload_stream(
                 {
                     resource_type: "video",
-                    folder: "opal"
+                    folder: "opal",
+                    public_id: data.filename,
                 },
-                (err,result) => {
+                async (err, result) => {
                     if (err) {
                         console.log("🔴 Error uploading file on cloudinary")
                     }
-                    console.log("🟢 Video uploaded to cloudinary : ",result.secure_url);
-                    fs.unlink("temp_upload/" + data.filename, (err) => {
-                        if (!err) {
-                            console.log(data.filename + " 🟢 deleted successfully");
-                        } else {
-                            console.log("🔴 Error deleting temporary file", err);
-                        }
-                    });
+                    console.log("🟢 Video uploaded to cloudinary : ", result.secure_url);
+                    const stopProcessing = await axios.post(
+                        `${processing.env.NEXT_API_HOST}recording/${data.userId}/complete`,
+                        { filename: data.filename }
+                    )
+                    if (stopProcessing.data.status !== 200) {
+                        console.log("🔴 Error: Something went wrong when stopping the process and trying to complete the processing stage.")
+                    }
+
+                    if (stopProcessing.data.status === 200) {
+                        fs.unlink("temp_upload/" + data.filename, (err) => {
+                            if (!err) console.log(data.filename + " " + "🟢 deleted successfuly")
+                        })
+                    }
                 }
             );
 
@@ -131,7 +143,7 @@ io.on("connection",(socket) => {
             //                             content: completion.choices[0].message.content,
             //                             transcript: transcription,
             //                         })
-                                    
+
             //                         if(titleSummaryGenerated.data.status !==200) {
             //                             console.log("🔴 Error: Something went wrong with creating the title and description")
             //                         }
@@ -161,11 +173,11 @@ io.on("connection",(socket) => {
 
         })
     })
-    socket.on("disconnect",async (data) =>  {
+    socket.on("disconnect", async (data) => {
         console.log("🔴 Disconnected from socket io")
     })
 })
-server.listen(5000,() => {
+server.listen(5000, () => {
     console.log("🟢 Listening on port 5000")
 })
